requirements.txt
flask-mail
pyjwt
werkzeug
wtforms
flask_login
flask-migrate
flask-sqlalchemy
flask-wtf
